finished the FLEX compiler folder
