+++ Flex/Executable/zeExecute.sh
@@ -16,7 +16,7 @@ do
   echo " "
   echo "EXECUTING" $item
   cd $item
-  ./Aprog $include/xin.txt $include/xout.txt $include/xlog.txt
+  ./Aprog < $include/xin.txt
   cd ..
 echo "EXECUTION COMPLETE"
 done
