+++ Flex/IOFiles/xin.txt
@@ -1,0 +1,1 @@
+When, in the course of human events, it becomes necessary for one people to dissolve the political bonds which have connected them with another, and to assume among the powers of the earth, the separate and equal station to which the laws of nature and of nature's God entitle them, a decent respect to the opinions of
