--- Lex/IOFiles/ErrorMSG.txt
@@ -1,13 +1,0 @@
-Executing zdCompile.sh
------------------------------------------------------------------
-Descend into 'TestDirectory' directory
- 
-COMPILING SourceFiles
-rm: cannot remove '*.o': No such file or directory
-rm: cannot remove 'a.out': No such file or directory
-lex: can't open main.lex
-cc: error: lex.yy.c: No such file or directory
-Return To 'WorkingDirectory' directory
- 
-COMPILING COMPLETE
-
