--- Lex/IOFiles/xin.txt
@@ -1,1 +1,0 @@
-xin.txt
