--- Lex/IOFiles/xlog.txt
@@ -1,26 +1,0 @@
-Skipping zaRefreshDate.sh
-Executing zbBackupProgram.sh
------------------------------------------------------------------
-Remove the old 'BackupDirectory' directory
-Create a new 'BackupDirectory' directory
-Descend into 'WorkingDirectory' directory
-Descend into directory SourceFiles
-Return from directory SourceFiles
-tar directory SourceFiles
-SourceFiles/
-SourceFiles/main.l
-gzip directory SourceFiles
-cp gzipped file to 'BackupDirectory'
-rm the gzipped file from  SourceFiles
-Return To 'WorkingDirectory' directory
-Executing zcCopyProgram.sh
------------------------------------------------------------------
-Remove the old 'TestDirectory' directory
-Create a new 'TestDirectory' directory
-cp files from 'BackupDirectory' into 'TestDirectory'
-Descend into 'TestDirectory' directory
-untar directory tarfileSourceFiles
-SourceFiles/
-SourceFiles/main.l
-Remove the tar files
-Return To 'WorkingDirectory' directory
